fix(wgpu): complete v30 migration across tests, examples and EPG crates
- wrap entry_point in Some and add cache: None to pipeline descriptors
- replace push/pop_error_scope with the v30 ErrorScopeGuard API
- handle fallible get_mapped_range/get_mapped_range_mut (MapRangeError)
- adapt to PollType polling and InstanceDescriptor without Default
- drop the removed request_device trace argument, add apply_limit_buckets
- map wgpu 30 u64 limit fields directly (removes useless conversions)
- resync Cargo.lock after dependency resolution

+++ src/wgpu_decode.rs
@@ -380,7 +380,7 @@ impl WgpuResidentDecodePipeline {
                 maximum,
             });
         }
-        let maximum_storage_bytes = u64::from(limits.max_storage_buffer_binding_size);
+        let maximum_storage_bytes = limits.max_storage_buffer_binding_size;
         validate_storage_binding_size("Q", layout.q_bytes, maximum_storage_bytes)?;
         validate_storage_binding_size("K", kv_bytes, maximum_storage_bytes)?;
         validate_storage_binding_size("V", kv_bytes, maximum_storage_bytes)?;
